requirements.txt
langchain 
langchain-community
pypdf
chromadb 
python-dotenv 
google-generativeai
streamlit
langchain 
langchain-community
pypdf
chromadb 
python-dotenv 
google-generativeai
streamlit
sentence-transformers
st-copy-to-clipboard
qdrant_client
streamlit-local-storage
langchain-text-splitters
pdfplumber
fastembed
numpy